feat: WR-724 use soft-delete for roadsegment in projections (#1015)
* feat: WR-724 use soft-delete for roadsegment in projections

* fix: update Begin fields when soft-deleting

+++ src/RoadRegistry.Wfs.Projections/RoadSegmentRecordProjection.cs
@@ -9,12 +9,13 @@ using BackOffice.Messages;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.Connector;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.SqlStreamStore;
 using Microsoft.EntityFrameworkCore;
+using RoadRegistry.BackOffice.FeatureToggles;
 using Schema;
 using Syndication.Schema;
 
 public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
 {
-    public RoadSegmentRecordProjection(IStreetNameCache streetNameCache)
+    public RoadSegmentRecordProjection(IStreetNameCache streetNameCache, UseRoadSegmentSoftDeleteFeatureToggle useRoadSegmentSoftDeleteFeatureToggle)
     {
         When<Envelope<SynchronizeWithStreetNameCache>>(async (context, envelope, token) =>
         {
@@ -105,7 +106,7 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
                         break;
 
                     case RoadSegmentRemoved roadSegmentRemoved:
-                        await RemoveRoadSegment(roadSegmentRemoved, context, token);
+                        await RemoveRoadSegment(roadSegmentRemoved, context, envelope, useRoadSegmentSoftDeleteFeatureToggle.FeatureEnabled, token);
                         break;
                 }
         });
@@ -131,10 +132,9 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
         var leftSideStreetNameRecord = await TryGetFromCache(streetNameCache, roadSegmentAdded.LeftSide.StreetNameId, token);
         var rightSideStreetNameRecord = await TryGetFromCache(streetNameCache, roadSegmentAdded.RightSide.StreetNameId, token);
 
-        await context.RoadSegments.AddAsync(new RoadSegmentRecord
+        await context.RoadSegments.AddAsync(UpdateBeginFields(new RoadSegmentRecord
         {
             Id = roadSegmentAdded.Id,
-            BeginTime = LocalDateTimeTranslator.TranslateFromWhen(envelope.Message.When),
             MaintainerId = roadSegmentAdded.MaintenanceAuthority.Code,
             MaintainerName = roadSegmentAdded.MaintenanceAuthority.Name,
             MethodDutchName = method.Translation.Name,
@@ -150,7 +150,7 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
             BeginRoadNodeId = roadSegmentAdded.StartNodeId,
             EndRoadNodeId = roadSegmentAdded.EndNodeId,
             StreetNameCachePosition = streetNameCachePosition
-        }, token);
+        }, envelope), token);
     }
 
     private static async Task ModifyRoadSegment(IStreetNameCache streetNameCache,
@@ -183,7 +183,6 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
         }
 
         roadSegmentRecord.Id = roadSegmentModified.Id;
-        roadSegmentRecord.BeginTime = LocalDateTimeTranslator.TranslateFromWhen(envelope.Message.When);
         roadSegmentRecord.MaintainerId = roadSegmentModified.MaintenanceAuthority.Code;
         roadSegmentRecord.MaintainerName = roadSegmentModified.MaintenanceAuthority.Name;
         roadSegmentRecord.MethodDutchName = method.Translation.Name;
@@ -199,6 +198,7 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
         roadSegmentRecord.BeginRoadNodeId = roadSegmentModified.StartNodeId;
         roadSegmentRecord.EndRoadNodeId = roadSegmentModified.EndNodeId;
         roadSegmentRecord.StreetNameCachePosition = streetNameCachePosition;
+        UpdateBeginFields(roadSegmentRecord, envelope);
     }
 
     private static async Task ModifyRoadSegmentAttributes(IStreetNameCache streetNameCache,
@@ -247,7 +247,7 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
             roadSegmentRecord.MaintainerName = roadSegmentAttributesModified.MaintenanceAuthority.Name;
         }
 
-        roadSegmentRecord.BeginTime = LocalDateTimeTranslator.TranslateFromWhen(envelope.Message.When);
+        UpdateBeginFields(roadSegmentRecord, envelope);
 
         var streetNameCachePosition = await streetNameCache.GetMaxPositionAsync(token);
         roadSegmentRecord.StreetNameCachePosition = streetNameCachePosition;
@@ -264,21 +264,36 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
         {
             throw new InvalidOperationException($"RoadSegmentRecord with id {segment.Id} is not found");
         }
-        
+
         roadSegmentRecord.Geometry2D = WfsGeometryTranslator.Translate2D(segment.Geometry);
 
-        roadSegmentRecord.BeginTime = LocalDateTimeTranslator.TranslateFromWhen(envelope.Message.When);
+        UpdateBeginFields(roadSegmentRecord, envelope);
 
         var streetNameCachePosition = await streetNameCache.GetMaxPositionAsync(token);
         roadSegmentRecord.StreetNameCachePosition = streetNameCachePosition;
     }
 
-    private static async Task RemoveRoadSegment(RoadSegmentRemoved roadSegmentRemoved, WfsContext context, CancellationToken token)
+    private static async Task RemoveRoadSegment(RoadSegmentRemoved roadSegmentRemoved,
+        WfsContext context,
+        Envelope<RoadNetworkChangesAccepted> envelope,
+        bool softDelete,
+        CancellationToken token)
     {
         var roadSegmentRecord = await context.RoadSegments.FindAsync(roadSegmentRemoved.Id, cancellationToken: token).ConfigureAwait(false);
         if (roadSegmentRecord is not null)
         {
-            context.RoadSegments.Remove(roadSegmentRecord);
+            if (softDelete)
+            {
+                if (!roadSegmentRecord.IsRemoved)
+                {
+                    UpdateBeginFields(roadSegmentRecord, envelope);
+                    roadSegmentRecord.IsRemoved = true;
+                }
+            }
+            else
+            {
+                context.RoadSegments.Remove(roadSegmentRecord);
+            }
         }
     }
 
@@ -288,5 +303,11 @@ public class RoadSegmentRecordProjection : ConnectedProjection<WfsContext>
         CancellationToken token)
     {
         return streetNameId.HasValue ? await streetNameCache.GetAsync(streetNameId.Value, token).ConfigureAwait(false) : null;
+    }
+
+    private static RoadSegmentRecord UpdateBeginFields(RoadSegmentRecord record, Envelope<RoadNetworkChangesAccepted> envelope)
+    {
+        record.BeginTime = LocalDateTimeTranslator.TranslateFromWhen(envelope.Message.When);
+        return record;
     }
 }
